requirements.txt
asgiref==3.7.2
Django==3.2
django-cors-headers==4.1.0
django-dotenv==1.4.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
Pillow==10.0.0
PyJWT==2.7.0
PyMySQL==1.1.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
uWSGI==2.0.21
